+++ src/main/php/peer/SocketInputStream.php
@@ -13,6 +13,7 @@ use stubbles\streams\InputStream;
  * Input stream for reading sockets.
  *
  * @api
+ * @deprecated  since 7.0.0, will be removed with 8.0.0
  */
 class SocketInputStream implements InputStream
 {