+++ src/main/php/peer/http/HttpRequest.php
@@ -9,7 +9,7 @@
  */
 namespace stubbles\peer\http;
 use stubbles\peer\HeaderList;
-use stubbles\streams\OutputStream;
+use stubbles\peer\Stream;
 /**
  * Class for sending a HTTP request.
  *
@@ -65,8 +65,8 @@ class HttpRequest
     public function get($timeout = 30, $version = HttpVersion::HTTP_1_1)
     {
         $socket = $this->httpUri->openSocket($timeout);
-        $this->processHeader($socket->out(), Http::GET, $version);
-        return HttpResponse::create($socket->in());
+        $this->processHeader($socket, Http::GET, $version);
+        return HttpResponse::create($socket);
     }
 
     /**
@@ -80,8 +80,8 @@ class HttpRequest
     {
         $socket = $this->httpUri->openSocket($timeout);
         $this->headers->put('Connection', 'close');
-        $this->processHeader($socket->out(), Http::HEAD, $version);
-        return HttpResponse::create($socket->in());
+        $this->processHeader($socket, Http::HEAD, $version);
+        return HttpResponse::create($socket);
     }
 
     /**
@@ -106,10 +106,9 @@ class HttpRequest
 
         $this->headers->put('Content-Length', strlen($body));
         $socket = $this->httpUri->openSocket($timeout);
-        $out    = $socket->out();
-        $this->processHeader($out, Http::POST, $version);
-        $out->write($body);
-        return HttpResponse::create($socket->in());
+        $this->processHeader($socket, Http::POST, $version);
+        $socket->write($body);
+        return HttpResponse::create($socket);
     }
 
     /**
@@ -125,10 +124,9 @@ class HttpRequest
     {
         $this->headers->put('Content-Length', strlen($body));
         $socket = $this->httpUri->openSocket($timeout);
-        $out    = $socket->out();
-        $this->processHeader($out, Http::PUT, $version);
-        $out->write($body);
-        return HttpResponse::create($socket->in());
+        $this->processHeader($socket, Http::PUT, $version);
+        $socket->write($body);
+        return HttpResponse::create($socket);
     }
 
     /**
@@ -142,8 +140,8 @@ class HttpRequest
     public function delete($timeout = 30, $version = HttpVersion::HTTP_1_1)
     {
         $socket = $this->httpUri->openSocket($timeout);
-        $this->processHeader($socket->out(), Http::DELETE, $version);
-        return HttpResponse::create($socket->in());
+        $this->processHeader($socket, Http::DELETE, $version);
+        return HttpResponse::create($socket);
     }
 
     /**
@@ -165,12 +163,12 @@ class HttpRequest
     /**
      * helper method to send the headers
      *
-     * @param   \stubbles\streams\OutputStream          $out      output stream to write request to
+     * @param   \stubbles\peer\Stream                   $socket      output stream to write request to
      * @param   string                                  $method   http method
      * @param   string|\stubbles\peer\http\HttpVersion  $version  http version
      * @throws  \InvalidArgumentException
      */
-    private function processHeader(OutputStream $out, $method, $version)
+    private function processHeader(Stream $socket, $method, $version)
     {
         $version = HttpVersion::castFrom($version);
         if (!$version->equals(HttpVersion::HTTP_1_0) && !$version->equals(HttpVersion::HTTP_1_1)) {
@@ -185,13 +183,13 @@ class HttpRequest
             $path .= '?' . $this->httpUri->queryString();
         }
 
-        $out->write(Http::line($method . ' ' . $path . ' ' . $version));
-        $out->write(Http::line('Host: ' . $this->httpUri->hostname()));
+        $socket->write(Http::line($method . ' ' . $path . ' ' . $version));
+        $socket->write(Http::line('Host: ' . $this->httpUri->hostname()));
         foreach ($this->headers as $key => $value) {
-            $out->write(Http::line($key . ': ' . $value));
+            $socket->write(Http::line($key . ': ' . $value));
         }
 
-        $out->write(Http::emptyLine());
+        $socket->write(Http::emptyLine());
     }
 
     /**