+++ src/test/php/peer/SocketInputStreamTest.php
@@ -75,7 +75,6 @@ class SocketInputStreamTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @test
-     * @group  foo
      */
     public function readLineFromSocketWithGivenLength()
     {